Move graph hooks into folder
